Fix sort_index overflow: int can't hold a millisecond timestamp
playlist_tracks.sort_index shipped as a 4-byte int (max ~2.1 billion).
Dragging a track onto a playlist writes Date.now() as its sort_index,
and a millisecond timestamp is already ~1.8 trillion today - nowhere
close to fitting. Every drop failed with Postgres's own "value ... is
out of range for type integer", confirmed live against the actual error.

The fix is the column, not the value: shrinking to something that
happens to fit int4 today (Unix seconds, say) just moves the same
overflow out to 2038 for no reason, when bigint holds a millisecond
timestamp forever with room to spare. playlists.sort_index is
untouched - it only ever holds a small array-length integer, never a
timestamp, so int was already the right type there.

schema.sql and 0003_playlists.sql both now say bigint directly, so
anyone provisioning fresh from here gets it right without ever seeing
this bug. 0004_playlist_tracks_bigint.sql is the catch-up ALTER for a
project that already ran the old 0003, like this one's live project
had.

+++ supabase/migrations/0003_playlists.sql
@@ -28,7 +28,7 @@ create table public.playlist_tracks (
   playlist_id  uuid not null references public.playlists(id) on delete cascade,
   track_id     uuid not null references public.tracks(id) on delete cascade,
   user_id      uuid not null default auth.uid() references auth.users(id) on delete cascade,
-  sort_index   int not null default 0,
+  sort_index   bigint not null default 0,
   added_at     timestamptz not null default now()
 );
